Small changes to eliminate warnings at compile.

+++ src/extras.c
@@ -125,7 +125,7 @@ static void posix_profil(caddr_t address, PAPI_sprofil_t * prof,
          - dividing by implicit 2 (2^^1 for a total of 2^^17), for even addresses
          NOTE: 131072 is a valid scale value. It produces byte resolution of addresses
       */
-      lloffset = ((size_t)(address - (size_t)prof->pr_off)) * prof->pr_scale;
+      lloffset = (address - prof->pr_off) * prof->pr_scale;
       indx = (unsigned long)(lloffset >> 17);
    }
